+++ NEWS.md
@@ -1,3 +1,7 @@
+# geomander 1.1.0
+* Collapse county calls into one for create_block_table()
+* Update variables to allow 2000 support in create_block_table()
+
 # geomander 1.0.9
 * Address SD and AK create_block_table assumptions